+++ src/apis/auth/steam.py
@@ -1,7 +1,6 @@
 # Copyright (C) 2022-2026 CharlesWithC All rights reserved.
 # Author: @CharlesWithC
 
-import json
 import time
 import uuid
 
@@ -55,7 +54,7 @@ async def get_callback(request: Request, response: Response):
         if app.config.steam_api_key != "":
             try:
                 r = await arequests.get(app, f"https://api.steampowered.com/ISteamUser/GetPlayerSummaries/v0002/?key={app.config.steam_api_key}&steamids={steamid}", dhrid = dhrid, timeout = 10)
-                d = json.loads(r.text)
+                d = r.json()
                 username = convertQuotation(d["response"]["players"][0]["personaname"])
                 avatar = convertQuotation(d["response"]["players"][0]["avatarfull"])
             except: