+++ changelog.md
@@ -3,6 +3,7 @@
 ## develop
 
 1. Improved Discord API response handling
+2. Replaced `json.loads(r.text)` with `r.json()` for more idiomatic json response handler
 
 ## v2.12.0